requirements.txt
flask==2.3.3
openai==1.3.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0